refactor: typescript files as requested by linter
