--- lib/components/map/connected-park-and-ride-overlay.js
@@ -1,46 +1,0 @@
-import ParkAndRideOverlay from '@opentripplanner/park-and-ride-overlay'
-import React, { Component } from 'react'
-import { connect } from 'react-redux'
-
-import { setLocation } from '../../actions/map'
-import { parkAndRideQuery } from '../../actions/api'
-
-class ConnectedParkAndRideOverlay extends Component {
-  componentDidMount () {
-    const params = {}
-    if (this.props.maxTransitDistance) {
-      params.maxTransitDistance = this.props.maxTransitDistance
-    }
-    // TODO: support config-defined bounding envelope
-
-    this.props.parkAndRideQuery(params)
-  }
-
-  render () {
-    return (
-      <ParkAndRideOverlay {...this.props} />
-    )
-  }
-}
-
-// connect to the redux store
-
-const mapStateToProps = (state, ownProps) => {
-  const { locations } = state.otp.overlay?.parkAndRide
-
-  // object type indicates error
-  if (typeof locations === 'object') {
-    return {}
-  }
-
-  return {
-    parkAndRideLocations: locations
-  }
-}
-
-const mapDispatchToProps = {
-  parkAndRideQuery,
-  setLocation
-}
-
-export default connect(mapStateToProps, mapDispatchToProps)(ConnectedParkAndRideOverlay)
