+++ lib/components/narrative/default/flex-indicator.js
@@ -11,6 +11,8 @@ import Icon from '../../util/icon'
 export const FLEX_COLOR = '#FA6400'
 const FLEX_COLOR_LIGHT = tinycolor(FLEX_COLOR).lighten(40).toHexString()
 
+// FIXME: type once the support-gtfs-flex branch is merged
+// eslint-disable-next-line react/prop-types
 const FlexNotice = ({ faKey, showText, text }) => (
   <>
     <Icon name={faKey} />
@@ -70,27 +72,28 @@ const FlexIndicatorWrapper = styled.div`
     width: 20px;
   }
 `
-
+// FIXME: type once the support-gtfs-flex branch is merged
+// eslint-disable-next-line react/prop-types
 export const FlexIndicator = ({ isCallAhead, isContinuousDropoff, shrink }) => (
   <FlexIndicatorWrapper shrink={shrink}>
     {!shrink && (
       <h4>
-        <FormattedMessage id='config.flex.flex-service' />
+        <FormattedMessage id="config.flex.flex-service" />
       </h4>
     )}
     {isCallAhead && (
       <FlexNotice
-        faKey='phone'
+        faKey="phone"
         showText={!shrink}
-        text={<FormattedMessage id='config.flex.call-ahead' />}
+        text={<FormattedMessage id="config.flex.call-ahead" />}
       />
     )}
     {/* Only show continuous dropoff message if call ahead message isn't shown */}
     {isContinuousDropoff && !isCallAhead && (
       <FlexNotice
-        faKey='hand-paper-o'
+        faKey="hand-paper-o"
         showText={!shrink}
-        text={<FormattedMessage id='config.flex.continuous-dropoff' />}
+        text={<FormattedMessage id="config.flex.continuous-dropoff" />}
       />
     )}
   </FlexIndicatorWrapper>
